Extract states, remove log
