+++ src/store/secondState.js
@@ -1,0 +1,53 @@
+const secondState = {
+  accountId: 2,
+  quality: {
+    category: 3,
+    score: 82,
+    details: [
+      {
+        name: 'Delivery time',
+        score: 18
+      },
+      {
+        name: 'Quality of products',
+        score: 7
+      },
+      {
+        name: 'Availability',
+        score: 20
+      }
+    ]
+  },
+  orders: [
+    {
+      name: 'AMD Ryzen 9 5950X',
+      buyer: 'Michael Green',
+      date: '2023-04-12',
+      price: '699',
+      id: 5
+    },
+    {
+      name: 'Intel Core i9-13900K',
+      buyer: 'Sarah White',
+      date: '2023-05-18',
+      price: '589',
+      id: 6
+    },
+    {
+      name: 'Apple M2 Max',
+      buyer: 'David Black',
+      date: '2023-06-25',
+      price: '1299',
+      id: 7
+    },
+    {
+      name: 'Samsung Galaxy Book3 Ultra',
+      buyer: 'Emily Blue',
+      date: '2023-07-30',
+      price: '2399',
+      id: 8
+    }
+  ]
+}
+
+export default secondState
